requirements.txt
djangorestframework>=3.15.2
Django>=5.1.1
psycopg2>=2.9.9